Initial work on moving deserialization into the event publishers
The serialized version of the event is inherently coupled with where the
event comes from, so push the deserialization into tthe publisher, which is
specific to the source of the event.
This allows the EventDispatcher to deal with deserialized events,
which makes it easier for it to deal with in-memory published events
that need no deserialization.