WIP - Add KurrentDb implementation
Add KurrentDb's experimental ARM image
Add extension method to convert to and from our Stream Version and KurrentDb's implementation.
Try to convert KurrentDb's WrongExpectedVersion exception to our internal version.
"Kurrently" one test fails because of attempting to read, either from the end of the stream => forwards, or from the start of the stream => backwards. Need to sort out that behaviour.